Finish snapshot
